+++ codebase-pipelines/codepipeline.tf
@@ -2,6 +2,7 @@ resource "aws_codepipeline" "codebase_pipeline" {
   for_each       = local.pipeline_map
   name           = "${var.application}-${var.codebase}-${each.value.name}-codebase-pipeline"
   role_arn       = aws_iam_role.codebase_deploy_pipeline.arn
+  depends_on     = [aws_iam_role_policy.artifact_store_access_for_codebase_pipeline]
   pipeline_type  = "V2"
   execution_mode = "QUEUED"
 
@@ -98,7 +99,7 @@ resource "aws_codepipeline" "codebase_pipeline" {
             ServiceName = "#{build_manifest.SERVICE_NAME_${upper(stage.value.name)}_${upper(replace(action.value.name, "-", "_"))}}"
             FileName    = "image-definitions-${action.value.name}.json"
           }
-          role_arn = "arn:aws:iam::${stage.value.account.id}:role/${var.application}-${stage.value.name}-codebase-pipeline-deploy-role"
+          role_arn = "arn:aws:iam::${stage.value.account.id}:role/${var.application}-${stage.value.name}-codebase-pipeline-deploy"
         }
       }
     }