+++ codebase-pipelines/eventbridge.tf
@@ -29,7 +29,7 @@ resource "aws_iam_role" "event_bridge_pipeline_trigger" {
 }
 
 resource "aws_iam_role_policy" "event_bridge_pipeline_trigger" {
-  name   = "${var.application}-${var.codebase}-pipeline-trigger-access-for-event-bridge"
+  name   = "event-bridge-access"
   role   = aws_iam_role.event_bridge_pipeline_trigger.name
   policy = data.aws_iam_policy_document.event_bridge_pipeline_trigger.json
 }