+++ codebase-pipelines/iam.tf
@@ -25,13 +25,13 @@ resource "aws_iam_role_policy_attachment" "ssm_access" {
   policy_arn = "arn:aws:iam::aws:policy/AmazonSSMReadOnlyAccess"
 }
 
-resource "aws_iam_role_policy" "log_access_for_codebase_image_build" {
-  name   = "${var.application}-${var.codebase}-log-access-for-codebase-pipeline-image-build"
+resource "aws_iam_role_policy" "log_access_for_codebuild_images" {
+  name   = "log-access"
   role   = aws_iam_role.codebase_image_build.name
-  policy = data.aws_iam_policy_document.log_access_for_codebuild.json
+  policy = data.aws_iam_policy_document.log_access.json
 }
 
-data "aws_iam_policy_document" "log_access_for_codebuild" {
+data "aws_iam_policy_document" "log_access" {
   statement {
     effect = "Allow"
     actions = [
@@ -41,48 +41,21 @@ data "aws_iam_policy_document" "log_access_for_codebuild" {
       "logs:TagLogGroup"
     ]
     resources = [
-      aws_cloudwatch_log_group.codebase_image_build.arn,
-      "${aws_cloudwatch_log_group.codebase_image_build.arn}:*",
-      "arn:aws:logs:${local.account_region}:log-group:*",
-      "arn:aws:codebuild:${local.account_region}:build/${var.application}-${var.codebase}-*-codebase-deploy-manifests",
-      "arn:aws:codebuild:${local.account_region}:build/${var.application}-${var.codebase}-*-codebase-deploy-manifests:*"
-    ]
-  }
-
-  statement {
-    effect = "Allow"
-    actions = [
-      "codebuild:CreateReportGroup",
-      "codebuild:CreateReport",
-      "codebuild:UpdateReport",
-      "codebuild:BatchPutTestCases",
-      "codebuild:BatchPutCodeCoverages"
-    ]
-    resources = [
-      "arn:aws:codebuild:${local.account_region}:report-group/${aws_codebuild_project.codebase_image_build.name}-*",
-      "arn:aws:codebuild:${local.account_region}:report-group/pipeline-${var.application}-*",
-      "arn:aws:codebuild:${local.account_region}:report-group/${var.application}-${var.codebase}-*-codebase-deploy-manifests-*"
+      "arn:aws:logs:${local.account_region}:log-group:codebuild/${var.application}-${var.codebase}-codebase-image-build/log-group",
+      "arn:aws:logs:${local.account_region}:log-group:codebuild/${var.application}-${var.codebase}-codebase-image-build/log-group:*",
+      "arn:aws:logs:${local.account_region}:log-group:codebuild/${var.application}-${var.codebase}-codebase-deploy-manifests/log-group",
+      "arn:aws:logs:${local.account_region}:log-group:codebuild/${var.application}-${var.codebase}-codebase-deploy-manifests/log-group:*"
     ]
   }
 }
 
-resource "aws_iam_role_policy" "ecr_access_for_codebase_image_build" {
-  name   = "${var.application}-${var.codebase}-ecr-access-for-codebase-pipeline-image-build"
+resource "aws_iam_role_policy" "ecr_access_for_codebuild_images" {
+  name   = "ecr-access"
   role   = aws_iam_role.codebase_image_build.name
-  policy = data.aws_iam_policy_document.ecr_access_for_codebase_image_build.json
+  policy = data.aws_iam_policy_document.ecr_access_for_codebuild_images.json
 }
 
-data "aws_iam_policy_document" "ecr_access_for_codebase_image_build" {
-  statement {
-    effect = "Allow"
-    actions = [
-      "ecr:GetAuthorizationToken"
-    ]
-    resources = [
-      "arn:aws:codebuild:${local.account_region}:report-group/pipeline-${var.application}-*"
-    ]
-  }
-
+data "aws_iam_policy_document" "ecr_access_for_codebuild_images" {
   statement {
     # checkov:skip=CKV_AWS_107:GetAuthorizationToken required for ci-image-builder
     effect = "Allow"
@@ -174,7 +147,7 @@ resource "aws_iam_role" "codebase_deploy_manifests" {
 }
 
 resource "aws_iam_role_policy" "artifact_store_access_for_codebuild_manifests" {
-  name   = "${var.application}-${var.codebase}-artifact-store-access-for-codebase-pipeline-deploy-manifests"
+  name   = "artifact-store-access"
   role   = aws_iam_role.codebase_deploy_manifests.name
   policy = data.aws_iam_policy_document.access_artifact_store.json
 }
@@ -223,13 +196,13 @@ data "aws_iam_policy_document" "access_artifact_store" {
 }
 
 resource "aws_iam_role_policy" "log_access_for_codebuild_manifests" {
-  name   = "${var.application}-${var.codebase}-log-access-for-codebase-pipeline-deploy-manifests"
+  name   = "log-access"
   role   = aws_iam_role.codebase_deploy_manifests.name
-  policy = data.aws_iam_policy_document.log_access_for_codebuild.json
+  policy = data.aws_iam_policy_document.log_access.json
 }
 
 resource "aws_iam_role_policy" "codebuild_assume_environment_deploy_role" {
-  name   = "${var.application}-${var.codebase}-environment-deploy-role-access-for-codebase-pipeline-deploy-manifests"
+  name   = "environment-deploy-role-access"
   role   = aws_iam_role.codebase_deploy_manifests.name
   policy = data.aws_iam_policy_document.assume_environment_deploy_role.json
 }
@@ -254,7 +227,7 @@ data "aws_iam_policy_document" "assume_codepipeline_role" {
 }
 
 resource "aws_iam_role_policy" "ecr_access_for_codebase_pipeline" {
-  name   = "${var.application}-${var.codebase}-ecr-access-for-codebase-pipeline"
+  name   = "ecr-access"
   role   = aws_iam_role.codebase_deploy_pipeline.name
   policy = data.aws_iam_policy_document.ecr_access_for_codebase_pipeline.json
 }
@@ -272,29 +245,26 @@ data "aws_iam_policy_document" "ecr_access_for_codebase_pipeline" {
 }
 
 resource "aws_iam_role_policy" "artifact_store_access_for_codebase_pipeline" {
-  name   = "${var.application}-${var.codebase}-artifact-store-access-for-codebase-pipeline"
+  name   = "artifact-store-access"
   role   = aws_iam_role.codebase_deploy_pipeline.name
   policy = data.aws_iam_policy_document.access_artifact_store.json
 }
 
 resource "aws_iam_role_policy" "pipeline_assume_environment_deploy_role" {
-  name   = "${var.application}-${var.codebase}-assume-environment-codebase-pipeline-deploy-role"
+  name   = "environment-deploy-role-access"
   role   = aws_iam_role.codebase_deploy_pipeline.name
   policy = data.aws_iam_policy_document.assume_environment_deploy_role.json
 }
 
 data "aws_iam_policy_document" "assume_environment_deploy_role" {
-  dynamic "statement" {
-    for_each = local.pipeline_environments
-
-    content {
-      effect = "Allow"
-      actions = [
-        "sts:AssumeRole"
-      ]
-      resources = [
-        "arn:aws:iam::${statement.value.account.id}:role/${var.application}-${statement.value.name}-codebase-pipeline-deploy-role"
-      ]
-    }
+  statement {
+    effect = "Allow"
+    actions = [
+      "sts:AssumeRole"
+    ]
+    resources = [
+      for env in local.pipeline_environments :
+      "arn:aws:iam::${env.account.id}:role/${var.application}-${env.name}-codebase-pipeline-deploy"
+    ]
   }
 }