+++ postgres/database-load/outputs.tf
@@ -1,3 +1,0 @@
-output "database-load-module-created" {
-  value = true
-}